refactor: 인터셉터 부분 코드 분리
